Add ExecuteInTransactionAsync to IEntityRepository
