+++ src/NuGetGallery.Core/Entities/EntityRepository.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -21,6 +22,21 @@ namespace NuGetGallery
         public async Task CommitChangesAsync()
         {
             await _entities.SaveChangesAsync();
+        }
+
+        /// <inheritdoc />
+        public async Task ExecuteInTransactionAsync(Func<Task> action)
+        {
+            action = action ?? throw new ArgumentNullException(nameof(action));
+
+#if NETFRAMEWORK
+            using (new SuspendDbExecutionStrategy())
+#endif
+            using (var transaction = _entities.GetDatabase().BeginTransaction())
+            {
+                await action();
+                transaction.Commit();
+            }
         }
 
         public void InsertOnCommit(T entity)
